requirements.txt
# --- For-Bot-Working --------- #
pyrofork==2.3.43
TgCrypto
pyromod==1.5
schedule
APscheduler
# --- For-Database ------------ #
pymongo
dnspython
# --- For-Web-Response ------- #
aiohttp
python-dotenv
# --- For-Web-Response ------- #
pytz
requests
bs4
shortzy
aiofiles
asyncio
